refactor: provide common hasCost function check
