+++ components/unit-cards/cost-card.tsx
@@ -1,27 +1,6 @@
 import { Flex, Image, Stack, Text, Title } from "@mantine/core";
 import { Fragment } from "react";
-
-/**
- * These fields can be found at `ebps` / `upgrade` inside each entity object.
- *
- * - For the units, look at `ebps` -> `exts` -> `template_reference` which gives
- *   the group within the Essence editor as `ebpextensions\\cost_ext`.
- * - For the upgrades, look at `upgrade` -> `upgrade_bag`.
- *
- * For both entities, the field is called `time_cost`.
- */
-export type ResourceValues = {
-  /** Value at `cost/fuel`. */
-  fuel?: number;
-  /** Value at `cost/munition`. */
-  munition?: number;
-  /** Value at `cost/munition`. */
-  manpower?: number;
-  /** Value at `cost/popcap`. */
-  popcap?: number;
-  /** Value at `time_seconds`. */
-  time_seconds?: number;
-};
+import { ResourceValues } from "../../src/unitStats";
 
 const ResourceIcons = [
   { icon: "/icons/common/resources/resource_buildtime_extra.png", type: "time_seconds" },
